[FLORA-378] Breadcrumbs for the package name and sub-resources
