+++ assets/css/1-core/2-variables.css
@@ -62,11 +62,12 @@
 
   /* --install-string-border: hsl(215 28% 17%); */
   --compiler-badge-background: hsl(218 12% 84%);
-  --text-color: hsl(221 39% 11%);
+  --text-color: var(--blue-20);
   --main-page-button-background: var(--gray-100);
   --main-page-button-focus-border-color: var(--purple-30);
   --main-page-button-divider: var(--purple-20);
   --deprecated-version: var(--red-60);
+  --breadcrumb-color: var(--gray-30);
 }
 
 html[data-theme="dark"] {
@@ -101,9 +102,10 @@ html[data-theme="dark"] {
 
   /* --install-string-border: hsl(215 28% 17%); */
   --compiler-badge-background: var(--blue-30);
-  --text-color: var(--blue-80);
+  --text-color: var(--gray-80);
   --main-page-button-background: var(--blue-20);
   --main-page-button-focus-border-color: var(--purple-30);
   --main-page-button-divider: var(--blue-15);
   --deprecated-version: var(--red-60);
+  --breadcrumb-color: hsl(221 9% 60%);
 }
