+++ assets/css/2-components/2-package-component.css
@@ -1,0 +1,19 @@
+summary.package-component {
+  cursor: pointer;
+  font-size: 1.25rem;
+  line-height: 1.25rem;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+  padding-left: 1rem;
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+}
+
+summary.package-component > * {
+  display: inline;
+}
+
+summary.package-component:hover {
+  background-color: var(--package-list-item-background-hover-color);
+  border-radius: 6px;
+}
