+++ assets/css/2-components/4-license.css
@@ -1,0 +1,9 @@
+/* stylelint-disable selector-class-pattern */
+/* stylelint-disable declaration-block-no-redundant-longhand-properties */
+
+.license-icon {
+  height: 1.25rem;
+  width: 1.25rem;
+  display: inline;
+  margin-top: -0.25rem;
+}
