+++ assets/css/3-screens/1-package/1-package.css
@@ -201,6 +201,14 @@
   }
 }
 
+a.breadcrumb-segment {
+  color: var(--breadcrumb-color);
+}
+
+a.breadcrumb-segment:hover {
+  text-decoration: underline;
+}
+
 @media (max-width: 47rem) {
   .package-body {
     grid-template-columns: repeat(1, 1fr);
