+++ assets/css/3-screens/5-package-version.css
@@ -1,0 +1,2 @@
+/* stylelint-disable selector-class-pattern */
+/* stylelint-disable declaration-block-no-redundant-longhand-properties */
