+++ assets/css/4-helpers.css
@@ -1,0 +1,22 @@
+/* stylelint-disable selector-class-pattern */
+/* stylelint-disable declaration-block-no-redundant-longhand-properties */
+
+.hidden {
+  display: none;
+}
+
+.font-light {
+  font-weight: 300;
+}
+
+.divider {
+  padding-bottom: 1rem;
+  margin-bottom: 1rem;
+  border-bottom: 1px solid hsl(0 0% 80%);
+}
+
+.divider-with-margin {
+  padding-bottom: 1rem;
+  margin-bottom: 4rem;
+  border-bottom: 1px solid hsl(0 0% 80%);
+}
