+++ design/package.json
@@ -21,6 +21,7 @@
     "@storybook/testing-library": "^0.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "storybook": "^7.2.0"
+    "storybook": "^7.2.0",
+    "vite": "^4.4.9"
   }
 }
